Add files via upload
Task 1: Re-write the program hangman, in your previous workshop (workshop – 2). The
program reads the words stored in a text file named hangman.txt. Words are delimited by
spaces.
Task 2:Write a program that prompts the user to enter a file name and displays the
occurrences of each letter in the file. Letters are case-insensitive. Here is a sample run:
Task 3: Standard telephone keypads contain the digits zero through nine. The numbers two
through nine each have three letters associated with them (figure below). Many people find it
difficult to memorize phone numbers, so they use the correspondence between digits and
letters to develop seven-letter words that correspond to their phone numbers.
For example, a person whose telephone number is 686-2377 might use the correspondence
Enter a filename: Somefile.txt
Number of A's: 56
Number of B's: 134
...
Number of Z's: 9
JAC – 444 Winter 2019
indicated in below figure to develop the seven-letter word “NUMBERS.” Every seven-letter
word corresponds to exactly one seven-digit telephone number. A restaurant wishing to
increase its takeout business could surely do so with the number 825-3688 (i.e., “TAKEOUT”).
Every seven-letter phone number corresponds to many different seven-letter words, but most
of these words represent unrecognizable juxtapositions of letters. It’s possible, however, that
the owner of a barbershop would be pleased to know that the shop’s telephone number, 424-
7288, corresponds to “HAIRCUT.” A veterinarian with the phone number 738-2273 would be
pleased to know that the number corresponds to the letters “PETCARE.” An automotive
dealership would be pleased to know that the dealership number, 639-2277, corresponds to
“NEWCARS.”
Your task is to write a program that, given a seven-digit number, write to a file every possible
seven-letter word combination corresponding to that number. There are 2,187 (3^7) such
combinations. Avoid phone numbers with the digits 0 and 1.